requirements.txt
streamlit
streamlit-feedback
openai
llama-index
nltk
python-dotenv